FREEPBX-13815 - Don't throw when pjsip is misconfigured
This ticket isn't SPECIFICALLY related to the bug, but, it's
mentioned in there.  The problem is that if someone has partly
configured a pjsip port, it'll return an empty string instead
of a number. This did throw an exception, but it's happened
enough times to just ignore the error. The protocol isn't
working, so there's nothing to unblock.