Fixed issue with PHP allowing access to /tmp
Previously, fixcustomrules script would create a tmp file to mark
that it has completed. The problem is that PHP save all /tmp files
to a private location. Basically, this means that any PHP that executes
will see a different folder for /tmp from the other scripts. Instead
of saving a file to /tmp, let's just check for the presence of the
files we're expecting to be created.