Check to see if IP exists in fail2ban ignore before adding
The existing code had the potential to add duplicate entries into fail2ban, causing difficulty in removing the ignore rule once a device registration would expire.  We now list all of the IPs in the system, and ensure that the IP we're trying to add is not in the list.