FREEPBX-22196 RFW Threshold settings UI Fixes
This commit addresses 2 changes in the RFW UI

* The new tunable RFW threshold settings are expanded
in the UI by default.  This looks a little messy, and
presents relatively advanced settings to the user,
inviting them to play with these settings when they
should really stay at default unless there's a problem.

* The warning banner advising caution against modifying
the tunable RFW settings is green, it should probably
be yellow like a warning.