FREEPBX-22563 Display Fail2Ban blocks on Attacks page
When the Sysadmin module is not activated,
IP addresses blocked by Fail2Ban are not able
to be manipulated within the GUI.  The RFW
seems to have been designed to work without
Fail2Ban in the way.

This commit will add the ability to view and
unban IP addresses that have been blocked by
the Fail2Ban asterisk-iptables jail when Sysadmin
has not been activated.  If Sysadmin is active,
the proper place to control this is from the
Intrusion Detection module.