FREEPBX-22574 Refresh RFW threshold rules
* RFW threshold parameters will update to the
database defined settings during rule refresh.
This removes the original logic in the initial
iteration of FREEPBX-22196 of querying the database
during the loading of default rules, and instead
includes this logic along with other dynamic aspects
of the firewall like registrations and RTP ports.

FREEPBX-22196 - Reorder ratelimit tiers

* The original iteration of this patch assigned
the tiers for rate limiting (1,2,3) in reverse order.
This commit will assign in the correct order.