FREEPBX-22682 Restore custom rules on firewall refresh
This should not be a common occurance, but if the fpbxfirewall
rule gets deleted somehow, and the default rules are reloaded,
the custom rules also need to be reloaded by calling
importCustomRules().

In order to prevent duplication of the custom rules on first startup,
the call to importCustomRules() needed to be removed from
updateFirewallRules and only be in checkFpbxFirewall.